+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Added
 - Added `Master Summoner` OP ability.
 - Added UMM option to allow Isekai dialogue when Isekai Protagonist is disabled.
+# Changed
+- Changed `Channel Energy` special powers to scale with Isekai Protagonist level rather than character level.
 # Fixed
 - Fixed all legacies which add features on class level.
 - Fixed Isekai dialogue showing up with Horgus when not Isekai Protagonist.
+- Fixed `Channel Negative Energy` not scaling with level.
 
 # [5.1.0] - 2023-05-07
 # Added